CODE/python-requirements.txt
connectorx==0.2.3
fsspec==2023.3.0
numpy==1.24.2
pandas==1.5.3
polars==0.16.12
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
black==23.1.0
isort==5.12.0
pre-commit==3.1.1
pytest==7.2.2
Flask-SQLAlchemy==3.0.3
pytest-flask==1.2.0
Flask-Cors==3.0.10
scikit-surprise==1.1.3
textblob==0.17.1
